refactor: 리뷰 반영 및 불필요한 파일 삭제 - issue #31
