+++ src/main/java/com/nexters/pinataserver/event/domain/EventRepositoryCustom.java
@@ -4,10 +4,10 @@ import java.util.List;
 
 import org.springframework.data.domain.Pageable;
 
-import com.nexters.pinataserver.event.dto.response.ParticipationEventDTO;
+import com.nexters.pinataserver.event.dto.response.ParticipationEventResponse;
 
 public interface EventRepositoryCustom {
 	List<Event> getMyEvents(Long userId, Pageable pageable);
 
-	List<ParticipationEventDTO> getParticipationEvents(Long userId, Pageable pageable);
+	List<ParticipationEventResponse> getParticipationEvents(Long userId, Pageable pageable);
 }
