+++ src/main/java/com/nexters/pinataserver/event/domain/EventStatus.java
@@ -1,7 +1,7 @@
 package com.nexters.pinataserver.event.domain;
 
 public enum EventStatus {
-	WAIT, PROCESS, COMPLETE, CANCEL;
+	PROCESS, WAIT, COMPLETE, CANCEL;
 
 	public boolean isComplete() {
 		return this == COMPLETE;
