+++ src/main/java/com/nexters/pinataserver/event/dto/response/ParticipateEventResponse.java
@@ -1,5 +1,8 @@
 package com.nexters.pinataserver.event.dto.response;
 
+import java.time.LocalDateTime;
+
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -37,6 +40,21 @@ public class ParticipateEventResponse {
 		this.itemId = itemId;
 		this.itemTitle = itemTitle;
 		this.itemImageUrl = itemImageUrl;
+	}
+
+	@Getter
+	@AllArgsConstructor
+	public class ParticipationEventDTO {
+		private Long eventId;
+		private String eventCode;
+		private boolean result;
+		private String resultMessage;
+		private String resultImageUrl;
+		private Long itemId;
+		private String itemImageUrl;
+		private boolean isPeriod;
+		private LocalDateTime openAt;
+		private LocalDateTime closeAt;
 	}
 
 }
