+++ src/main/java/com/nexters/pinataserver/event/service/EventReadService.java
@@ -11,7 +11,7 @@ import com.nexters.pinataserver.common.exception.e4xx.NotFoundException;
 import com.nexters.pinataserver.common.util.ImageUtil;
 import com.nexters.pinataserver.event.domain.Event;
 import com.nexters.pinataserver.event.domain.EventRepository;
-import com.nexters.pinataserver.event.dto.response.EventResponse;
+import com.nexters.pinataserver.event.dto.response.OrganizersEventResponse;
 import com.nexters.pinataserver.event.dto.response.ParticipationEventResponse;
 import com.nexters.pinataserver.event.dto.response.ReadCurrentParticipateEventResponse;
 
@@ -63,17 +63,14 @@ public class EventReadService {
 			.build();
 	}
 
-	public List<EventResponse> getEvents(Long userId, Pageable pageable) {
+	public List<OrganizersEventResponse> getEvents(Long userId, Pageable pageable) {
 		return eventRepository.getMyEvents(userId, pageable).stream()
-			.map(EventResponse::from)
+			.map(OrganizersEventResponse::from)
 			.collect(Collectors.toList());
 	}
 
 	public List<ParticipationEventResponse> getParticipationEvents(Long userId, Pageable pageable) {
-		return eventRepository.getParticipationEvents(userId, pageable)
-			.stream()
-			.map(ParticipationEventResponse::from)
-			.collect(Collectors.toList());
+		return eventRepository.getParticipationEvents(userId, pageable);
 	}
 
 }
